add snyk source class
